# Conflicts:
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/furu.f90
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/setextforcechkadvec.f90

+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute/links_to_centers.f90
@@ -41,7 +41,7 @@ module m_links_to_centers
       module procedure links_to_centers_dp_rank_2
    end interface links_to_centers
 contains
-   !> Set flow node value based on flow link values, where vlin is double precision
+   !> Set flow node value based on flow link values, where vlin is real(kind=dp)
    subroutine links_to_centers_dp(vnod, vlin)
       use precision, only: dp
       use m_flow, only: lnkx, ndkx, kmx, kmxn
